"new reference data for gdgem clustering; updates #155"
