+++ prototype2/gdgem/clustering/test_data/config.json
@@ -1,0 +1,58 @@
+{
+  "builder_type" : "VMM3",
+
+  "time_config" :
+  {
+    "tac_slope" : 100,
+    "bc_clock" : 20,
+    "trigger_resolution" : 1,
+    "target_resolution" : 1,
+    "acquisition_window" : 8191
+  },
+
+  "srs_mappings" :
+  [
+    {"fecID":1, "vmmID":0, "planeID":0, "strip_offset":0},
+    {"fecID":1, "vmmID":1, "planeID":0, "strip_offset":64},
+    {"fecID":1, "vmmID":14, "planeID":1, "strip_offset":0},
+    {"fecID":1, "vmmID":15, "planeID":1, "strip_offset":64}
+  ],
+
+  "clusterer x" :
+  {
+    "hit_adc_threshold" : 0,
+    "max_strip_gap" : 0,
+    "max_time_gap" : 200,
+    "min_cluster_size" : 3
+  },
+
+  "clusterer y" :
+  {
+    "hit_adc_threshold" : 0,
+    "max_strip_gap" : 0,
+    "max_time_gap" : 200,
+    "min_cluster_size" : 3
+  },
+
+  "matcher_max_delta_time" : 500,
+
+  "analyze_weighted" : true,
+  "analyze_max_timebins" : 3,
+  "analyze_max_timedif" : 7,
+
+  "filters" :
+  {
+    "enforce_lower_uncertainty_limit" : false,
+    "lower_uncertainty_limit" : 6,
+    "enforce_minimum_hits" : false,
+    "minimum_hits" : 6
+  },
+
+  "hit_histograms" : true,
+  "cluster_adc_downshift" : 6,
+  "send_tracks" : true,
+  "track_sample_minhits" : 1,
+
+  "geometry_x" : 384,
+  "geometry_y" : 384
+}
